VANGUARD-Phase-1/pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "accessibility-auditor"
version = "0.1.0"
description = "AI-Powered Accessibility Gap Detection System for Urban Public Transport"
readme = "README.md"
requires-python = ">=3.9"
license = {text = "MIT"}
authors = [
    {name = "Accessibility Team", email = "team@example.com"}
]

dependencies = [
    "fastapi>=0.104.0",
    "uvicorn>=0.24.0",
    "pydantic>=2.0.0",
    "pydantic-settings>=2.0.0",
    "pandas>=2.0.0",
    "numpy>=1.24.0",
    "scikit-learn>=1.3.0",
    "nltk>=3.8.0",
    "spacy>=3.7.0",
    "sentence-transformers>=2.2.0",
    "torch>=2.0.0",
    "ultralytics>=8.0.0",
    "streamlit>=1.28.0",
    "folium>=0.14.0",
    "plotly>=5.17.0",
    "reportlab>=4.0.0",
    "weasyprint>=59.0",
    "Pillow>=10.0.0",
    "python-multipart>=0.0.6",
    "aiofiles>=23.2.0",
    "python-dateutil>=2.8.0",
    "pytest>=7.4.0",
    "pytest-cov>=4.1.0",
    "pytest-asyncio>=0.21.0",
    "black>=23.9.0",
    "flake8>=6.1.0",
    "pylint>=3.0.0",
]

[project.optional-dependencies]
dev = [
    "jupyter>=1.0.0",
    "ipython>=8.16.0",
]

[tool.setuptools]
packages = ["src"]

[tool.black]
line-length = 100
target-version = ["py39", "py310", "py311"]

[tool.pylint."messages control"]
disable = ["C0111", "R0913"]

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
addopts = "--cov=src --cov-report=html --cov-report=term-missing"
